Create VISION.md
With the recent uptake in interest in Reken, it is time to write down the vision I have for Reken. I started Reken 3 years ago after being frustrated with all the tooling (and dependencies) that were required to get a web framework up and running. In a startup I was working on at the time decided to go all Vanilla javascript but then stranded in the complexities of all the dom manipulation code. The vision for Reken evolved from these experiences.